drm: color pipeline base work
This is a squashed patch based on a series sent out by Harry wentland.
It contains all the changes in the series(v7) currently under review
below.

https://patchwork.freedesktop.org/series/123446/#rev7

This patch lays the ground work for incremental changes and Intel
specific pipeline changes.

NOTE: This patch is not meant for review. Any review related to this
patch should be done on the original series. In order not to diverge
the discussion from the main series.